# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Smart ERP API
  version: 1.1.0
  summary: Machine form of docs/spec/04-api-contracts.md
  description: |
    Seeded from `docs/spec/04-api-contracts.md` version 1.1.0. That file is the human form and
    wins on conflict until this document is regenerated. Additive changes bump the minor version;
    breaking changes bump the major version and carry a migration note.

    1.1.0 (additive): `GET /me/sessions` and `DELETE /me/sessions/{id}` plus schema `UserSession`
    (R1.15, A9). Period soft-close, hard-close, audit-adjustment open, and the holiday calendar.
    No breaking change from 1.0.0.

    Transport rules that apply to every operation:

    - JSON over HTTPS. All timestamps are UTC ISO 8601. Money is `{ amount, currency }` with the
      amount as a decimal string.
    - Every success response is `{ data, meta }`. Every failure, including 429 and gateway errors,
      is `{ error: { code, message, details, request_id } }`.
    - Every mutating request accepts `Idempotency-Key` (UUID); the server replays a stored response
      for 24 hours; a different body with the same key is `409 CONFLICT`.
    - Every mutable resource carries `state_version`; mutating requests send `If-Match`; a mismatch
      is `409 CONFLICT` with the current state in `error.details`.
    - `Accept-Language: en | ar` affects messages and print formats, never data.
  contact:
    name: Smart ERP contract owner
  license:
    name: Proprietary
...
servers:
...
security:
  - bearer: []
    dpop: []
...
paths:
...
  /sod-matrix:
    get:
      operationId: listSodMatrix
      tags: [authz]
      summary: List the segregation-of-duties matrix
      description: |
        Incompatible role pairs and incompatible actions on one document (R1.13). Includes
        pending rows created by a mutation that is waiting on an approval request (04 Masters).
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "200":
          description: Matrix rows visible to the caller
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SodRule"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
    post:
      operationId: createSodRule
      tags: [authz]
      summary: Propose a segregation-of-duties rule
      description: |
        Sensitive master mutation (04 Masters). The row is stored as `pending_approval` and is
        not enforced until the approval request is decided. `If-Match` carries `state_version`
        (`0` when creating). An incompatible role assignment is refused with `SOD_VIOLATION`
        unless an override approval is cited (A14); that enforcement is on role assignment, and
        this route is the matrix itself (A23 when the approval engine consumes the request).
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SodRuleWrite"
      responses:
        "200":
          description: Rule stored, pending approval or already decided with the current row
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/SodRule"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /approval-matrix:
    get:
      operationId: listApprovalMatrix
      tags: [authz]
      summary: List approval matrix rules
      description: |
        Per document type: amount threshold, below-threshold role, independent first approver
        and final gate (R2.2). Pending mutations are included.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "200":
          description: Matrix rows visible to the caller
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApprovalMatrixRule"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
    post:
      operationId: createApprovalMatrixRule
      tags: [authz]
      summary: Propose an approval matrix rule
      description: |
        Sensitive master mutation (04 Masters). Stored as `pending_approval` until the approval
        engine decides the request. `If-Match` carries `state_version` (`0` when creating).
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalMatrixRuleWrite"
      responses:
        "200":
          description: Rule stored pending approval
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/ApprovalMatrixRule"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /holiday-calendar:
    get:
      operationId: getHolidayCalendar
      tags: [masters]
      summary: Read the company holiday calendar
      description: |
        Weekend definition, business-hours window, and holidays used by every business-hours
        clock (R13.8, C5, ADR-12).
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "200":
          description: Calendar
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/HolidayCalendar"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
    post:
      operationId: putHolidayCalendar
      tags: [masters]
      summary: Create or replace the company holiday calendar
      description: |
        Replaces weekend days, the business-hours window, and the holiday list in one write.
        `If-Match` carries the current `state_version`; the first write sends `0` (R13.8, A19).
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatchRequired"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HolidayCalendarWrite"
      responses:
        "200":
          description: Calendar saved
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/HolidayCalendar"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: ERP access token issued by POST /auth/token. Expires in 15 minutes or less.
    dpop:
      type: apiKey
      in: header
      name: DPoP
      description: |
        DPoP proof JWT signed by the enrolled device key (RFC 9449). Required on every mobile
        request alongside the bearer token; the console uses a session cookie with a CSRF token.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        UUID chosen by the client. The server stores the response for 24 hours and replays it for a
        repeated key with the same body; a different body with the same key is `409 CONFLICT`.
      schema:
        type: string
        format: uuid
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: |
        The `state_version` the client last saw. A mismatch is `409 CONFLICT` with the current
        state in `error.details.current`.
      schema:
        type: string
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: "`en` or `ar`. Affects messages and print formats, never data."
      schema:
        type: string
        enum: [en, ar]
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from `meta.next_cursor` of the previous page.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    IfMatchRequired:
      name: If-Match
      in: header
      required: true
      description: |
        The `state_version` the client last saw. Required on period and holiday-calendar
        writes. A mismatch is `409 CONFLICT` with the current state in `error.details.current`.
        The first holiday-calendar write sends `0`.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Validation failed (`VALIDATION_ERROR`) or a required header is missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: "`AUTH_REQUIRED`, `TOKEN_EXPIRED`, `TOKEN_REVOKED` or `DEVICE_MISMATCH`"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: "`PERMISSION_DENIED`, `STEP_UP_REQUIRED` or `SOD_VIOLATION`"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: "`NOT_FOUND`"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: |
        `CONFLICT` (stale `If-Match`, current state in `error.details.current`, or a reused
        `Idempotency-Key` with a different body), `IMMUTABLE` or `PERIOD_CLOSED`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: "`RATE_LIMITED`, in the standard envelope (A7, A17)"
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the client may retry
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: "`INTERNAL_ERROR`; the message never leaks internals, `request_id` correlates logs"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unavailable:
      description: "`INTERNAL_ERROR` or `MISSING_CONFIG` while the service is not ready; also returned by the gateway in the same envelope"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Money:
      type: object
      description: Decimal amount as a string plus an ISO 4217 currency code. Never a float.
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: "^-?[0-9]+(\\.[0-9]+)?$"
          examples: ["1234.56"]
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          examples: ["AED"]
    Meta:
      type: object
      description: Present on every success response.
      required: [as_of, request_id]
      properties:
        as_of:
          type: string
          format: date-time
          description: Server time the response was computed, UTC
        request_id:
          type: string
          description: Correlates the response with logs and audit rows
      additionalProperties: true
    PageMeta:
      description: Meta for list responses. `total` is present only where cheap.
      allOf:
        - $ref: "#/components/schemas/Meta"
        - type: object
          properties:
            next_cursor:
              type: [string, "null"]
            total:
              type: integer
              minimum: 0
...
    ErrorCode:
      type: string
      description: The closed set of error codes from docs/spec/04-api-contracts.md.
      enum:
        - AUTH_REQUIRED
        - TOKEN_EXPIRED
        - TOKEN_REVOKED
        - DEVICE_MISMATCH
        - STEP_UP_REQUIRED
        - PERMISSION_DENIED
        - VALIDATION_ERROR
        - MISSING_CONFIG
        - NOT_FOUND
        - CONFLICT
        - ALREADY_DECIDED
        - IMMUTABLE
        - PERIOD_CLOSED
        - SOD_VIOLATION
        - CREDIT_HOLD
        - PRICE_FLOOR_HOLD
        - NEGATIVE_STOCK
        - RATE_LIMITED
        - INTERNAL_ERROR
    Error:
      type: object
      required: [code, message, details, request_id]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Human readable, localised by Accept-Language
        details:
          type: object
          description: Code-specific detail, e.g. `{ field }` for VALIDATION_ERROR or `{ current }` for CONFLICT
          additionalProperties: true
        request_id:
          type: string
    ErrorEnvelope:
      type: object
      description: Returned on every failure path including 429 and gateway errors (A17).
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    StateVersion:
      type: integer
      minimum: 0
      description: Optimistic concurrency version; bumped on every transition
...
    SodRule:
      type: object
      description: One row of the segregation-of-duties matrix (R1.13).
      required: [id, kind, left_code, right_code, state_version, status]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [role_pair, action_pair]
          description: |
            `role_pair` is two roles one person must not hold together. `action_pair` is two
            actions one person must not both perform on a document (enter and approve, receive
            and count, create vendor and pay vendor, request correction and approve correction).
        left_code:
          type: string
          description: Role name or permission code
        right_code:
          type: string
          description: Role name or permission code
        state_version:
          $ref: "#/components/schemas/StateVersion"
        status:
          type: string
          enum: [active, pending_approval]
        approval_request_id:
          type: [string, "null"]
          description: Set when the mutation is waiting on the approval engine
    SodRuleWrite:
      type: object
      required: [kind, left_code, right_code]
      properties:
        kind:
          type: string
          enum: [role_pair, action_pair]
        left_code:
          type: string
          minLength: 1
        right_code:
          type: string
          minLength: 1
    ApprovalMatrixRule:
      type: object
      description: Approval routing for one document type (R2.2). The engine that consumes it is the approvals module.
      required:
        - id
        - document_type
        - threshold
        - below_threshold_role
        - first_approver_role
        - final_gate_role
        - state_version
        - status
      properties:
        id:
          type: string
          format: uuid
        document_type:
          type: string
        threshold:
          $ref: "#/components/schemas/Money"
          description: At or above this amount the first approver and the final gate both apply
        below_threshold_role:
          type: string
          description: Any one holder of this role may register the document below the threshold
        first_approver_role:
          type: string
          description: Independent first approver at or above the threshold
        final_gate_role:
          type: string
          description: Named final gate; not delegable
        voting_any:
          type: [integer, "null"]
          minimum: 1
          description: When set with `voting_of`, the company requires any N of M
        voting_of:
          type: [integer, "null"]
          minimum: 1
        state_version:
          $ref: "#/components/schemas/StateVersion"
        status:
          type: string
          enum: [active, pending_approval]
        approval_request_id:
          type: [string, "null"]
    ApprovalMatrixRuleWrite:
      type: object
      required: [document_type, threshold, below_threshold_role, first_approver_role, final_gate_role]
      properties:
        document_type:
          type: string
          minLength: 1
        threshold:
          $ref: "#/components/schemas/Money"
        below_threshold_role:
          type: string
          minLength: 1
        first_approver_role:
          type: string
          minLength: 1
        final_gate_role:
          type: string
          minLength: 1
        voting_any:
          type: [integer, "null"]
          minimum: 1
        voting_of:
          type: [integer, "null"]
          minimum: 1
...
    Weekday:
      type: string
      enum: [sunday, monday, tuesday, wednesday, thursday, friday, saturday]
    Holiday:
      type: object
      required: [date, name]
      additionalProperties: false
      properties:
        date:
          type: string
          format: date
        name:
          type: string
          minLength: 1
    HolidayCalendar:
      type: object
      description: |
        Company business-hours calendar (ADR-12). Clocks count only time inside
        `business_open`–`business_close` on days that are neither weekend nor holiday.
      required: [state_version, timezone, business_open, business_close, weekend, holidays]
      properties:
        state_version:
          $ref: "#/components/schemas/StateVersion"
        timezone:
          type: string
          minLength: 1
          examples: ["Asia/Dubai"]
        business_open:
          type: string
          description: Local start of the working day, HH:MM (24-hour)
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
          examples: ["08:00"]
        business_close:
          type: string
          description: Local end of the working day, HH:MM (24-hour), after `business_open`
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
          examples: ["20:00"]
        weekend:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/Weekday"
        holidays:
          type: array
          items:
            $ref: "#/components/schemas/Holiday"
    HolidayCalendarWrite:
      type: object
      description: Body of POST /holiday-calendar. `state_version` is not accepted from the client.
      required: [timezone, business_open, business_close, weekend, holidays]
      additionalProperties: false
      properties:
        timezone:
          type: string
          minLength: 1
        business_open:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
        business_close:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
        weekend:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/Weekday"
        holidays:
          type: array
          items:
            $ref: "#/components/schemas/Holiday"
...